moving Nav to component
