+++ src/pages/home/subscriptions/index.js
@@ -1,7 +1,7 @@
 import React, { useEffect, useGlobal, useState, useRef } from "reactn";
 import styled from "styled-components";
 import { Plans } from "./Plans";
-import { Navbar } from "../Navbar";
+import { Navbar } from "../../../components/Navbar";
 
 export const Subscriptions = (props) => {
   const [authUser] = useGlobal("user");
